pyproject.toml
[build-system]
requires = [
    "setuptools>=73.0.1",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "fandango-fuzzer"
version = "0.8.0"
authors = [
    { name = "José Antonio Zamudio Amaya", email = "jose.zamudio@cispa.de" },
    { name = "Marius Smytzek", email = "marius.smytzek@cispa.de" },
    { name = "Andreas Zeller", email = "andreas.zeller@cispa.de" },
]
description = "Fandango produces myriads of high-quality random inputs to test programs, giving users unprecedented control over format and shape of the inputs."
readme = "README.md"
license = { file = "LICENSE.md" }
requires-python = ">=3.10,<4.0"
classifiers = [
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: European Union Public Licence 1.2 (EUPL 1.2)",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Testing"
]
dependencies = [
    "pytest>=7.2.2",
    "antlr4-python3-runtime>=4.13",
    "deprecation>=2.1.0",
    "ansi_styles>=0.2.2",
    "cachedir_tag>=0.0.3",
    "xdg-base-dirs>=6.0.2",
    "dill>=0.3.7",
    "thefuzz>=0.22",
    "py010parser>=0.1.18",
    "exrex>=0.12.0",
    "gnureadline>=8.2.13 ; platform_system != 'Windows'",
    "pyreadline3>=3.5.4; platform_system == 'Windows'",
    "tdigest>=0.5.2.2",
    "tqdm>=4.67.1",
    "PyYAML>=6.0.2",
    "cachetools>=5.5.2",
]

[project.urls]
homepage = "https://fandango-fuzzer.github.io/"
repository = "https://github.com/fandango-fuzzer/fandango/"
"Bug Tracker" = "https://github.com/fandango-fuzzer/fandango/issues"

[project.scripts]
fandango = "fandango.cli:main"

[project.optional-dependencies]
test = [
    "pytest-cov>=4.1.0",
    "pytest-html>=3.2.0",
    "pytest-rerunfailures>=11.1.2",
    "parameterized>=0.8.1",
    "Faker>=30.4.0",
    "docutils>=0.20.1",  # Jupyter-book needs docutils<=0.20
    "tccbox",
    "python-dateutil>=2.9.0.post0",
    "deap>=1.4.1",
    "scipy>=1.14.1",
    "numpy>=2.2.1",  # Avoid conflict with pandas
    "matplotlib>=3.9.2",
    "ordered-set>=4.0.2",
    "cryptography>=44.0.1"
]

[tool.black]
line-length = 88
target-version = ['py311']
